Move proxy logic into a separate module
