+++ scrapinghub/client/requests.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import
 
-from .utils import _Proxy
+from .proxy import _Proxy
 
 
 class Requests(_Proxy):
